Keep the world clock and its stamps as doubles
A float clock loses resolution as a session runs: two hours in, its
step is a millisecond, and the senses' one-update-old check on a shot
stamp fails at about four. The clock and the four stamps taken from it
are doubles now. Steps stay float, since a step is small, and so does
an age from secondsSince. A double fed float steps cannot overflow, so
that guard and its test are gone. A test stamps the clock five hours in
and still measures one step.

+++ include/simple_platformer/actor/actor.hpp
@@ -39,7 +39,7 @@ namespace simple_platformer
 
         LifeState life = LifeState::Alive;
         float deathTimeRemaining = 0.0F;
-        std::optional<float> lastDamageTimeSeconds;
+        std::optional<double> lastDamageTimeSeconds;
 
         std::optional<Sprite> sprite;
         std::optional<Animator> animator;
